% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phitest.R
\name{phitest}
\alias{phitest}
\title{The generic function for the Phitest method}
\usage{
phitest(object, ...)
}
\arguments{
\item{object}{An object of single-cell data}

\item{...}{Other parameters to be passed to the function.}
}
\value{
A list of two elements: \code{pval} contains the \emph{P} values, and
\code{par} contains the estimated parameters.
}
\description{
Phitest for analyzing the heterogeneity of single-cell populations
}
\author{
Wei Vivian Li, \email{vivian.li@rutgers.edu}
}
