% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phitest.R
\name{phitest.Seurat}
\alias{phitest.Seurat}
\title{Appling the Phitest method with a Seurat object}
\usage{
\method{phitest}{Seurat}(object, ncores = 1, min.cell = 10)
}
\arguments{
\item{object}{A Seurat object.}

\item{ncores}{Number of cores used for parallel computation.}

\item{min.cell}{An integer specifying a threshold to filter genes. Genes expressed in
fewer than \code{min.cell} cells are filtered out.}
}
\value{
A list of two elements: \code{pval} contains the \emph{P} values, and
\code{par} contains the estimated parameters.
}
\description{
Phitest for analyzing the heterogeneity of single-cell populations
}
\author{
Wei Vivian Li, \email{vivian.li@rutgers.edu}
}
